Added UserProfile persistence
Same UserProfile now persists across all pages of the app. More than one user can be created, and it is easy to switch between different users.

NOTE: All users / info are deleted upon closing of the app. I'm planning to work on storing this information in a text file or other database solution in order to prevent this going forward.

SIDE NOTE: commented out the 'onCreateOptionsMenu' method in UserProfilePageActivity.java (was causing the app to crash when ran)